requirements.txt
SQLAlchemy
PyMySQL
mysql-connector-python
psycopg2-binary
pandas
streamlit
python-dotenv
kagglehub
openai
google-generativeai
sqlparse
sqlglot
openpyxl
xlsxwriter
tabulate
fastapi
uvicorn
python-multipart